Move identifier records into separate directory
