+++ Dfe.Identifiers.Api/Controllers/IdentifierController.cs
@@ -1,6 +1,7 @@
 using System.Text.Json;
 using System.Text.RegularExpressions;
 using Dfe.Identifiers.Api.Models;
+using Dfe.Identifiers.Api.Models.Identifiers;
 using Dfe.Identifiers.Api.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -15,12 +16,12 @@ namespace Dfe.Identifiers.Api.Controllers;
 [SwaggerTag("Identifiers Endpoints")]
 public class IdentifiersController : ControllerBase
 {
-    private readonly TrustRepository _trustQueries;
-    private readonly EstablishmentRepository _establishmentQueries;
+    private readonly ITrustRepository _trustQueries;
+    private readonly IEstablishmentRepository _establishmentQueries;
     private readonly ILogger<IdentifiersController> _logger;
 
-    public IdentifiersController(TrustRepository trustQueries, ILogger<IdentifiersController> logger,
-        EstablishmentRepository establishmentQueries)
+    public IdentifiersController(ITrustRepository trustQueries, ILogger<IdentifiersController> logger,
+        IEstablishmentRepository establishmentQueries)
     {
         _trustQueries = trustQueries;
         _logger = logger;
@@ -38,7 +39,7 @@ public class IdentifiersController : ControllerBase
     [SwaggerOperation(Summary = "Retrieves an object's Identifiers based on one of its identifiers.",
         Description = "Returns an objects identifiers found in the database.")]
     [SwaggerResponse(200, "Successfully found and returned the objects identifiers.")]
-    public async Task<ActionResult<Identifiers>> GetIdentifiers(string identifier, CancellationToken cancellationToken)
+    public async Task<ActionResult<Models.Identifiers.Identifiers>> GetIdentifiers(string identifier, CancellationToken cancellationToken)
     {
         var loggableIdentifier =
             Regex.Replace(identifier, "[^a-zA-Z0-9-]", "", RegexOptions.None, TimeSpan.FromSeconds(2));
@@ -46,13 +47,13 @@ public class IdentifiersController : ControllerBase
         var trusts = await _trustQueries.GetTrustsByIdentifier(identifier, cancellationToken).ConfigureAwait(false);
         var establishments = await _establishmentQueries.GetEstablishmentsByIdentifier(identifier, cancellationToken)
             .ConfigureAwait(false);
-        var results = new Identifiers(trusts.Select(MapToIdentifiers).ToArray(), establishments.Select(MapToIdentifiers).ToArray());
+        var results = new Models.Identifiers.Identifiers(trusts.Select(MapToIdentifiers).ToArray(), establishments.Select(MapToIdentifiers).ToArray());
         _logger.LogInformation("Returning objects found by identifier {identifier}", loggableIdentifier);
         _logger.LogDebug("{output}", JsonSerializer.Serialize(results));
         return Ok(results);
     }
 
-    public static EstablishmentIdentifiers MapToIdentifiers(Establishment establishment)
+    private static EstablishmentIdentifiers MapToIdentifiers(Establishment establishment)
     {
         return new EstablishmentIdentifiers(UKPRN: establishment.UKPRN, URN: establishment.URN.ToString(),
             LAESTAB: $"{establishment.LocalAuthority.Code}/{establishment.EstablishmentNumber}");
@@ -63,17 +64,3 @@ public class IdentifiersController : ControllerBase
         return new TrustIdentifiers(UID: trust.GroupUID, UKPRN: trust.UKPRN, TrustReference: trust.GroupID);
     }
 }
-
-public record Identifiers(TrustIdentifiers[] Trusts, EstablishmentIdentifiers[] Establishments);
-
-public record TrustIdentifiers(
-    string? UID,
-    string? UKPRN,
-    string? TrustReference
-);
-
-public record EstablishmentIdentifiers(
-    string? LAESTAB,
-    string? UKPRN,
-    string? URN
-);
