+++ Dfe.Identifiers.Api/Models/Identifiers/EstablishmentIdentifiers.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfe.Identifiers.Api.Models.Identifiers;
+
+public record EstablishmentIdentifiers(
+    string? LAESTAB,
+    string? UKPRN,
+    string? URN
+);
