+++ Dfe.Identifiers.Api/Models/Identifiers/Identifiers.cs
@@ -1,0 +1,3 @@
+﻿namespace Dfe.Identifiers.Api.Models.Identifiers;
+
+public record Identifiers(TrustIdentifiers[] Trusts, EstablishmentIdentifiers[] Establishments);
