+++ Dfe.Identifiers.Api/Models/Identifiers/TrustIdentifiers.cs
@@ -1,0 +1,7 @@
+﻿namespace Dfe.Identifiers.Api.Models.Identifiers;
+
+public record TrustIdentifiers(
+    string? UID,
+    string? UKPRN,
+    string? TrustReference
+);
